Merge in SV/sv from feature/SVN-2296-Support_for_gcc13 to develop

* commit 'de134863ef0a865448ca3f201bd70b2f4d91e41d':
  workaround to provide minimal support for development on mingw
  consistently search for Python in all CMake projects
  do not link with both static and dynamic version of secp256k1 library
  fix errors/warnings reported by gcc13

+++ src/validation.cpp
@@ -3494,7 +3494,7 @@ private:
         size_t maxThreads { parallelTxnValidation? static_cast<size_t>(config.GetPerBlockTxnValidatorThreadsCount()) : 1UL };
         uint64_t batchSize { config.GetBlockValidationTxBatchSize() };
         size_t numThreads { block.vtx.size() / batchSize };
-        numThreads = std::clamp(numThreads, 1UL, maxThreads);
+        numThreads = std::clamp(numThreads, size_t(1), maxThreads);
 
         int64_t startGroupTime { GetTimeMicros() };
         TxnGrouper grouper {};
